refactor(backend): move struct to where it makes sense
